Move WS tests to separate module
